add: alt to skills images
